Make `event-cpi` feature opt-in instead of opt-out
