+++ lang/syn/Cargo.toml
@@ -16,7 +16,7 @@ hash = []
 default = []
 anchor-debug = []
 seeds = []
-cpi-events = []
+event-cpi = []
 
 [dependencies]
 proc-macro2 = { version = "1.0", features=["span-locations"]}
