+++ lang/syn/src/codegen/program/dispatch.rs
@@ -27,9 +27,13 @@ pub fn generate(program: &Program) -> proc_macro2::TokenStream {
             }
         })
         .collect();
+
     let fallback_fn = gen_fallback(program).unwrap_or(quote! {
         Err(anchor_lang::error::ErrorCode::InstructionFallbackNotFound.into())
     });
+
+    let event_cpi_handler = generate_event_cpi_handler();
+
     quote! {
         /// Performs method dispatch.
         ///
@@ -83,21 +87,7 @@ pub fn generate(program: &Program) -> proc_macro2::TokenStream {
                     }
                 }
                 anchor_lang::event::EVENT_IX_TAG_LE => {
-                   // If the method identifier is the event tag, then execute an event cpi
-                   // against the noop instruction injected into all Anchor programs unless they have
-                   // no-cpi-events enabled.
-                    #[cfg(not(feature = "no-cpi-events"))]
-                    {
-                        __private::__events::__event_dispatch(
-                            program_id,
-                            accounts,
-                            &ix_data,
-                        )
-                    }
-                    #[cfg(feature = "no-cpi-events")]
-                    {
-                        Err(anchor_lang::error::ErrorCode::EventInstructionStub.into())
-                    }
+                    #event_cpi_handler
                 }
                 _ => {
                     #fallback_fn
@@ -116,4 +106,18 @@ pub fn gen_fallback(program: &Program) -> Option<proc_macro2::TokenStream> {
             #program_name::#fn_name(program_id, accounts, data)
         }
     })
+}
+
+/// Generate the event-cpi instruction handler based on whether the `event-cpi` feature is enabled.
+pub fn generate_event_cpi_handler() -> proc_macro2::TokenStream {
+    #[cfg(feature = "event-cpi")]
+    quote! {
+        // `event-cpi` feature is enabled, dispatch self-cpi instruction
+        __private::__events::__event_dispatch(program_id, accounts, &ix_data)
+    }
+    #[cfg(not(feature = "event-cpi"))]
+    quote! {
+        // `event-cpi` feature is not enabled
+        Err(anchor_lang::error::ErrorCode::EventInstructionStub.into())
+    }
 }
