+++ tests/events/programs/events/Cargo.toml
@@ -14,7 +14,6 @@ no-entrypoint = []
 no-idl = []
 cpi = ["no-entrypoint"]
 default = []
-no-cpi-events = []
 
 [dependencies]
-anchor-lang = { path = "../../../../lang" }
+anchor-lang = { path = "../../../../lang", features = ["event-cpi"] }
